Final push 17/11

+++ database/migrations/2022_11_13_134339_create_housings_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->text('descrizione');
             $table->string('costo')->nullable();
             $table->string('city');
-            $table->integer('numero_telefono')->nullable();
+            $table->bigInteger('numero_telefono')->nullable();
             $table->timestamps();
         });
     }
